Using this lib makes elm, deno, hotreload without vite. But the error message is broken
