+++ public/index.html
@@ -31,10 +31,6 @@
 
   <body>
     <div id="mobu"></div>
-    <script src="main.js">
-      debugger
-    </script>
-
     <script type="module" src="bridge.js"></script>
   </body>
 </html>
