+++ scripts/compile.ts
@@ -1,0 +1,3 @@
+import { compileToModuleString } from 'https://deno.land/x/deno_elm_compiler@0.1.0/compiler.ts';
+
+console.log(await compileToModuleString('./src/Main.elm'));
